move amsi analysis library to mumfim
This commit is the first step in bringing in the amsi analysis library
directly into mumfim. This analysis library is currently tightly
integrated with the code in the macroscale solver. We are working to
decople the class definitions. We found that having the analysis library
in a separate package made the necessary refactorings difficult.
Additionally, no external dependencies rely on AMSI, so this shouldn't
be problematic.

Further, we want to increase the separation of amsi as the "parallel
control" library and mumfim as the analysis package.